renderer: Split up renderer.cpp
