+++ src/emulator/renderer/src/renderer.cpp
@@ -40,41 +40,6 @@ namespace renderer {
         return source;
     }
 
-    static SharedGLObject compile_glsl(GLenum type, const std::string &source) {
-        R_PROFILE(__func__);
-
-        const SharedGLObject shader = std::make_shared<GLObject>();
-        if (!shader->init(glCreateShader(type), glDeleteShader)) {
-            return SharedGLObject();
-        }
-
-        const GLchar *source_glchar = static_cast<const GLchar *>(source.c_str());
-        const GLint length = static_cast<GLint>(source.length());
-        glShaderSource(shader->get(), 1, &source_glchar, &length);
-
-        glCompileShader(shader->get());
-
-        GLint log_length = 0;
-        glGetShaderiv(shader->get(), GL_INFO_LOG_LENGTH, &log_length);
-
-        if (log_length > 0) {
-            std::vector<GLchar> log;
-            log.resize(log_length);
-            glGetShaderInfoLog(shader->get(), log_length, nullptr, log.data());
-
-            LOG_ERROR("{}", log.data());
-        }
-
-        GLboolean is_compiled = GL_FALSE;
-        glGetShaderiv(shader->get(), GL_COMPILE_STATUS, &is_compiled);
-        assert(is_compiled != GL_FALSE);
-        if (!is_compiled) {
-            return SharedGLObject();
-        }
-
-        return shader;
-    }
-
     static SceGxmParameterType parameter_type(const SceGxmProgramParameter &parameter) {
         return static_cast<SceGxmParameterType>(static_cast<uint16_t>(parameter.type));
     }
@@ -445,14 +410,6 @@ namespace renderer {
         return locations;
     }
 
-    static void bind_attribute_locations(GLuint gl_program, const VertexProgram &program) {
-        R_PROFILE(__func__);
-
-        for (const AttributeLocations::value_type &binding : program.attribute_locations) {
-            glBindAttribLocation(gl_program, binding.first / sizeof(uint32_t), binding.second.c_str());
-        }
-    }
-
     static void flip_vertically(uint32_t *pixels, size_t width, size_t height, size_t stride_in_pixels) {
         R_PROFILE(__func__);
 
@@ -463,318 +420,6 @@ namespace renderer {
             std::swap_ranges(&row1[0], &row1[width], &row2[0]);
             row1 += stride_in_pixels;
             row2 -= stride_in_pixels;
-        }
-    }
-
-    static SharedGLObject get_program(Context &context, const GxmContextState &state, const MemState &mem) {
-        R_PROFILE(__func__);
-
-        assert(state.fragment_program);
-        assert(state.vertex_program);
-
-        const FragmentProgram &fragment_program = *state.fragment_program.get(mem)->renderer.get();
-        const VertexProgram &vertex_program = *state.vertex_program.get(mem)->renderer.get();
-        const ProgramGLSLs glsls(fragment_program.glsl, vertex_program.glsl);
-        const ProgramCache::const_iterator cached = context.program_cache.find(glsls);
-        if (cached != context.program_cache.end()) {
-            return cached->second;
-        }
-
-        const SharedGLObject fragment_shader = compile_glsl(GL_FRAGMENT_SHADER, fragment_program.glsl);
-        if (!fragment_shader) {
-            LOG_CRITICAL("Error in compiled fragment shader:\n{}", fragment_program.glsl);
-            return SharedGLObject();
-        }
-        const SharedGLObject vertex_shader = compile_glsl(GL_VERTEX_SHADER, vertex_program.glsl);
-        if (!vertex_shader) {
-            LOG_CRITICAL("Error in compiled vertex shader:\n{}", vertex_program.glsl);
-            return SharedGLObject();
-        }
-
-        const SharedGLObject program = std::make_shared<GLObject>();
-        if (!program->init(glCreateProgram(), &glDeleteProgram)) {
-            return SharedGLObject();
-        }
-
-        glAttachShader(program->get(), fragment_shader->get());
-        glAttachShader(program->get(), vertex_shader->get());
-
-        bind_attribute_locations(program->get(), vertex_program);
-
-        glLinkProgram(program->get());
-
-        GLint log_length = 0;
-        glGetProgramiv(program->get(), GL_INFO_LOG_LENGTH, &log_length);
-
-        if (log_length > 0) {
-            std::vector<GLchar> log;
-            log.resize(log_length);
-            glGetProgramInfoLog(program->get(), log_length, nullptr, log.data());
-
-            LOG_ERROR("{}\n", log.data());
-        }
-
-        GLboolean is_linked = GL_FALSE;
-        glGetProgramiv(program->get(), GL_LINK_STATUS, &is_linked);
-        assert(is_linked != GL_FALSE);
-        if (is_linked == GL_FALSE) {
-            return SharedGLObject();
-        }
-
-        glDetachShader(program->get(), fragment_shader->get());
-        glDetachShader(program->get(), vertex_shader->get());
-
-        context.program_cache.emplace(glsls, program);
-
-        return program;
-    }
-
-    template <class T>
-    static void uniform_4(GLint location, GLsizei count, const T *value);
-
-    template <>
-    void uniform_4<GLfloat>(GLint location, GLsizei count, const GLfloat *value) {
-        glUniform4fv(location, count, value);
-    }
-
-    template <>
-    void uniform_4<GLint>(GLint location, GLsizei count, const GLint *value) {
-        glUniform4iv(location, count, value);
-    }
-
-    template <class T>
-    static void uniform_1(GLint location, GLsizei count, const T *value);
-
-    template <>
-    void uniform_1<GLfloat>(GLint location, GLsizei count, const GLfloat *value) {
-        glUniform1fv(location, count, value);
-    }
-
-    template <>
-    void uniform_1<GLint>(GLint location, GLsizei count, const GLint *value) {
-        glUniform1iv(location, count, value);
-    }
-
-    template <class T>
-    static void uniform_matrix_4(GLint location, GLsizei count, GLboolean, const T *value);
-
-    template <>
-    void uniform_matrix_4<GLfloat>(GLint location, GLsizei count, GLboolean transpose, const GLfloat *value) {
-        glUniformMatrix4fv(location, count, transpose, value);
-    }
-
-    template <>
-    void uniform_matrix_4<GLint>(GLint location, GLsizei count, GLboolean transpose, const GLint *value) {
-        glUniform4iv(location, count, (GLint *)value);
-    }
-
-    template <class T>
-    static void set_uniform(GLint location, size_t component_count, GLsizei array_size, const T *value) {
-        switch (component_count) {
-        case 1:
-            switch (array_size) {
-            case 1:
-                uniform_1<T>(location, array_size, value);
-                break;
-            }
-            break;
-        case 4:
-            if (array_size % 4 == 0) {
-                uniform_matrix_4<T>(location, array_size / 4, GL_FALSE, value);
-            } else {
-                uniform_4<T>(location, array_size, value);
-            }
-            break;
-
-        default:
-            LOG_WARN("Unhandled uniform component count {}.", component_count);
-            break;
-        }
-    }
-
-    static void set_uniforms(GLuint gl_program, const UniformBuffers &uniform_buffers, const SceGxmProgram &gxm_program, const MemState &mem) {
-        R_PROFILE(__func__);
-
-        const SceGxmProgramParameter *const parameters = gxm::program_parameters(gxm_program);
-        for (size_t i = 0; i < gxm_program.parameter_count; ++i) {
-            const SceGxmProgramParameter &parameter = parameters[i];
-            if (parameter.category != SCE_GXM_PARAMETER_CATEGORY_UNIFORM) {
-                continue;
-            }
-
-            auto name = gxm::parameter_name(parameter);
-            const GLint location = glGetUniformLocation(gl_program, name.c_str());
-            if (location < 0) {
-                LOG_WARN("Uniform parameter {} not found in current OpenGL program.", name);
-                continue;
-            }
-
-            const SceGxmParameterType type = static_cast<SceGxmParameterType>(static_cast<uint16_t>(parameter.type));
-            const Ptr<const void> uniform_buffer = uniform_buffers[parameter.container_index];
-            if (!uniform_buffer) {
-                LOG_WARN("Uniform buffer {} not set for parameter {}.", parameter.container_index, name);
-                continue;
-            }
-
-            const GLint *src_s32;
-            const GLfloat *src_f32;
-
-            const uint8_t *const base = static_cast<const uint8_t *>(uniform_buffer.get(mem));
-            switch (type) {
-            case SCE_GXM_PARAMETER_TYPE_S32:
-                src_s32 = reinterpret_cast<const GLint *>(base + parameter.resource_index * 4); // TODO What offset?
-                set_uniform<GLint>(location, parameter.component_count, parameter.array_size, src_s32);
-                break;
-            case SCE_GXM_PARAMETER_TYPE_F32:
-                src_f32 = reinterpret_cast<const GLfloat *>(base + parameter.resource_index * 4); // TODO What offset?
-                set_uniform<GLfloat>(location, parameter.component_count, parameter.array_size, src_f32);
-                break;
-            default:
-                LOG_WARN("Type {} not handled for uniform parameter {}.", type, name);
-                break;
-            }
-        }
-    }
-
-    static void set_uniforms(GLuint program, const Context &context, const GxmContextState &state, const MemState &mem) {
-        R_PROFILE(__func__);
-
-        assert(state.fragment_program);
-        assert(state.vertex_program);
-
-        const SceGxmFragmentProgram &fragment_program = *state.fragment_program.get(mem);
-        const SceGxmVertexProgram &vertex_program = *state.vertex_program.get(mem);
-        assert(fragment_program.program);
-        assert(vertex_program.program);
-
-        set_uniforms(program, state.fragment_uniform_buffers, *fragment_program.program.get(mem), mem);
-        set_uniforms(program, state.vertex_uniform_buffers, *vertex_program.program.get(mem), mem);
-    }
-
-    static GLenum translate_depth_func(SceGxmDepthFunc depth_func) {
-        R_PROFILE(__func__);
-
-        switch (depth_func) {
-        case SCE_GXM_DEPTH_FUNC_NEVER:
-            return GL_NEVER;
-        case SCE_GXM_DEPTH_FUNC_LESS:
-            return GL_LESS;
-        case SCE_GXM_DEPTH_FUNC_EQUAL:
-            return GL_EQUAL;
-        case SCE_GXM_DEPTH_FUNC_LESS_EQUAL:
-            return GL_LEQUAL;
-        case SCE_GXM_DEPTH_FUNC_GREATER:
-            return GL_GREATER;
-        case SCE_GXM_DEPTH_FUNC_NOT_EQUAL:
-            return GL_NOTEQUAL;
-        case SCE_GXM_DEPTH_FUNC_GREATER_EQUAL:
-            return GL_GEQUAL;
-        case SCE_GXM_DEPTH_FUNC_ALWAYS:
-            return GL_ALWAYS;
-        }
-
-        return GL_ALWAYS;
-    }
-
-    static GLenum translate_stencil_op(SceGxmStencilOp stencil_op) {
-        R_PROFILE(__func__);
-
-        switch (stencil_op) {
-        case SCE_GXM_STENCIL_OP_KEEP:
-            return GL_KEEP;
-        case SCE_GXM_STENCIL_OP_ZERO:
-            return GL_ZERO;
-        case SCE_GXM_STENCIL_OP_REPLACE:
-            return GL_REPLACE;
-        case SCE_GXM_STENCIL_OP_INCR:
-            return GL_INCR;
-        case SCE_GXM_STENCIL_OP_DECR:
-            return GL_DECR;
-        case SCE_GXM_STENCIL_OP_INVERT:
-            return GL_INVERT;
-        case SCE_GXM_STENCIL_OP_INCR_WRAP:
-            return GL_INCR_WRAP;
-        case SCE_GXM_STENCIL_OP_DECR_WRAP:
-            return GL_DECR_WRAP;
-        }
-
-        return GL_KEEP;
-    }
-
-    static GLenum translate_stencil_func(SceGxmStencilFunc stencil_func) {
-        R_PROFILE(__func__);
-
-        switch (stencil_func) {
-        case SCE_GXM_STENCIL_FUNC_NEVER:
-            return GL_NEVER;
-        case SCE_GXM_STENCIL_FUNC_LESS:
-            return GL_LESS;
-        case SCE_GXM_STENCIL_FUNC_EQUAL:
-            return GL_EQUAL;
-        case SCE_GXM_STENCIL_FUNC_LESS_EQUAL:
-            return GL_LEQUAL;
-        case SCE_GXM_STENCIL_FUNC_GREATER:
-            return GL_GREATER;
-        case SCE_GXM_STENCIL_FUNC_NOT_EQUAL:
-            return GL_NOTEQUAL;
-        case SCE_GXM_STENCIL_FUNC_GREATER_EQUAL:
-            return GL_GEQUAL;
-        case SCE_GXM_STENCIL_FUNC_ALWAYS:
-            return GL_ALWAYS;
-        }
-
-        return GL_ALWAYS;
-    }
-
-    static void apply_stencil_state(GLenum face, const GxmStencilState &state) {
-        glStencilOpSeparate(face,
-            translate_stencil_op(state.stencil_fail),
-            translate_stencil_op(state.depth_fail),
-            translate_stencil_op(state.depth_pass));
-        glStencilFuncSeparate(face, translate_stencil_func(state.func), state.ref, state.compare_mask);
-        glStencilMaskSeparate(face, state.write_mask);
-    }
-
-    static GLenum attribute_format_to_gl_type(SceGxmAttributeFormat format) {
-        R_PROFILE(__func__);
-
-        switch (format) {
-        case SCE_GXM_ATTRIBUTE_FORMAT_U8:
-        case SCE_GXM_ATTRIBUTE_FORMAT_U8N:
-            return GL_UNSIGNED_BYTE;
-        case SCE_GXM_ATTRIBUTE_FORMAT_S8:
-        case SCE_GXM_ATTRIBUTE_FORMAT_S8N:
-            return GL_BYTE;
-        case SCE_GXM_ATTRIBUTE_FORMAT_U16:
-        case SCE_GXM_ATTRIBUTE_FORMAT_U16N:
-            return GL_UNSIGNED_SHORT;
-        case SCE_GXM_ATTRIBUTE_FORMAT_S16:
-        case SCE_GXM_ATTRIBUTE_FORMAT_S16N:
-            return GL_SHORT;
-        case SCE_GXM_ATTRIBUTE_FORMAT_F16:
-            return GL_HALF_FLOAT;
-        case SCE_GXM_ATTRIBUTE_FORMAT_F32:
-            return GL_FLOAT;
-        case SCE_GXM_ATTRIBUTE_FORMAT_UNTYPED:
-            LOG_WARN("Unsupported attribute format SCE_GXM_ATTRIBUTE_FORMAT_UNTYPED");
-            return GL_UNSIGNED_BYTE;
-        }
-
-        LOG_ERROR("Unsupported attribute format {}", log_hex(format));
-        return GL_UNSIGNED_BYTE;
-    }
-
-    static GLboolean attribute_format_normalised(SceGxmAttributeFormat format) {
-        R_PROFILE(__func__);
-
-        switch (format) {
-        case SCE_GXM_ATTRIBUTE_FORMAT_U8N:
-        case SCE_GXM_ATTRIBUTE_FORMAT_S8N:
-        case SCE_GXM_ATTRIBUTE_FORMAT_U16N:
-        case SCE_GXM_ATTRIBUTE_FORMAT_S16N:
-            return GL_TRUE;
-        default:
-            return GL_FALSE;
         }
     }
 
@@ -880,132 +525,6 @@ namespace renderer {
         ++context.texture_cache.timestamp;
 
         glBindFramebuffer(GL_FRAMEBUFFER, 0);
-    }
-
-    bool sync_state(Context &context, const GxmContextState &state, const MemState &mem, bool enable_texture_cache) {
-        R_PROFILE(__func__);
-
-        // TODO Use some kind of caching to avoid setting every draw call?
-        const SharedGLObject program = get_program(context, state, mem);
-        if (!program) {
-            return false;
-        }
-        glUseProgram(program->get());
-
-        // Viewport.
-        const GLsizei display_w = state.color_surface.pbeEmitWords[0];
-        const GLsizei display_h = state.color_surface.pbeEmitWords[1];
-        const GxmViewport &viewport = state.viewport;
-        if (viewport.enable == SCE_GXM_VIEWPORT_ENABLED) {
-            const GLfloat w = viewport.scale.x * 2;
-            const GLfloat h = viewport.scale.y * -2;
-            const GLfloat x = viewport.offset.x - viewport.scale.x;
-            const GLfloat y = display_h - (viewport.offset.y - viewport.scale.y);
-            glViewportIndexedf(0, x, y, w, h);
-            glDepthRange(viewport.offset.z - viewport.scale.z, viewport.offset.z + viewport.scale.z);
-        } else {
-            glViewport(0, 0, display_w, display_h);
-            glDepthRange(0, 1);
-        }
-
-        // Clipping.
-        // TODO: There was some math here to round the scissor rect, but it looked potentially incorrect.
-        const GLsizei scissor_x = state.region_clip_min.x;
-        const GLsizei scissor_y = display_h - state.region_clip_min.y;
-        const GLsizei scissor_w = state.region_clip_max.x - state.region_clip_min.x;
-        const GLsizei scissor_h = state.region_clip_max.y - state.region_clip_min.y;
-        switch (state.region_clip_mode) {
-        case SCE_GXM_REGION_CLIP_NONE:
-            glDisable(GL_SCISSOR_TEST);
-            break;
-        case SCE_GXM_REGION_CLIP_ALL:
-            glEnable(GL_SCISSOR_TEST);
-            glScissor(0, 0, 0, 0);
-            break;
-        case SCE_GXM_REGION_CLIP_OUTSIDE:
-            glEnable(GL_SCISSOR_TEST);
-            glScissor(scissor_x, scissor_y, scissor_w, scissor_h);
-            break;
-        case SCE_GXM_REGION_CLIP_INSIDE:
-            // TODO: Implement SCE_GXM_REGION_CLIP_INSIDE
-            glDisable(GL_SCISSOR_TEST);
-            LOG_WARN("Unimplemented region clip mode used: SCE_GXM_REGION_CLIP_INSIDE");
-            break;
-        }
-
-        // Culling.
-        switch (state.cull_mode) {
-        case SCE_GXM_CULL_CCW:
-            glEnable(GL_CULL_FACE);
-            glCullFace(GL_FRONT);
-            break;
-        case SCE_GXM_CULL_CW:
-            glEnable(GL_CULL_FACE);
-            glCullFace(GL_BACK);
-            break;
-        case SCE_GXM_CULL_NONE:
-            glDisable(GL_CULL_FACE);
-            break;
-        }
-
-        // Depth test.
-        if (state.depth_stencil_surface.depthData) {
-            glEnable(GL_DEPTH_TEST);
-            glDepthFunc(translate_depth_func(state.front_depth_func));
-            glDepthMask(state.front_depth_write_enable == SCE_GXM_DEPTH_WRITE_ENABLED ? GL_TRUE : GL_FALSE);
-        } else {
-            glDisable(GL_DEPTH_TEST);
-        }
-
-        // Stencil.
-        if (state.depth_stencil_surface.stencilData) {
-            glEnable(GL_STENCIL_TEST);
-            apply_stencil_state(GL_FRONT, state.front_stencil);
-            apply_stencil_state(GL_BACK, state.back_stencil);
-        } else {
-            glDisable(GL_STENCIL_TEST);
-        }
-
-        // Blending.
-        const FragmentProgram &fragment_program = *state.fragment_program.get(mem)->renderer.get();
-        glColorMask(fragment_program.color_mask_red, fragment_program.color_mask_green, fragment_program.color_mask_blue, fragment_program.color_mask_alpha);
-        if (fragment_program.blend_enabled) {
-            glEnable(GL_BLEND);
-            glBlendEquationSeparate(fragment_program.color_func, fragment_program.alpha_func);
-            glBlendFuncSeparate(fragment_program.color_src, fragment_program.color_dst, fragment_program.alpha_src, fragment_program.alpha_dst);
-        } else {
-            glDisable(GL_BLEND);
-        }
-
-        // Textures.
-        for (size_t i = 0; i < SCE_GXM_MAX_TEXTURE_UNITS; ++i) {
-            const SceGxmTexture &texture = state.fragment_textures[i];
-            if (texture.data_addr != 0) {
-                glActiveTexture(static_cast<GLenum>(GL_TEXTURE0 + i));
-                cache_and_bind_texture(context.texture_cache, texture, mem, enable_texture_cache);
-            }
-        }
-        glActiveTexture(GL_TEXTURE0);
-
-        // Uniforms.
-        set_uniforms(program->get(), context, state, mem);
-
-        // Vertex attributes.
-        const SceGxmVertexProgram &vertex_program = *state.vertex_program.get(mem);
-        for (const emu::SceGxmVertexAttribute &attribute : vertex_program.attributes) {
-            const SceGxmVertexStream &stream = vertex_program.streams[attribute.streamIndex];
-            const SceGxmAttributeFormat attribute_format = static_cast<SceGxmAttributeFormat>(attribute.format);
-            const GLenum type = attribute_format_to_gl_type(attribute_format);
-            const GLboolean normalised = attribute_format_normalised(attribute_format);
-            const int attrib_location = attribute.regIndex / sizeof(uint32_t);
-
-            glBindBuffer(GL_ARRAY_BUFFER, context.stream_vertex_buffers[attribute.streamIndex]);
-            glVertexAttribPointer(attrib_location, attribute.componentCount, type, normalised, stream.stride, reinterpret_cast<const GLvoid *>(attribute.offset));
-            glEnableVertexAttribArray(attrib_location);
-        }
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-        return true;
     }
 
     void finish(Context &context) {
